compiles and passes tests, proly something wrong with it

+++ src/org/adligo/i/log/client/I_LogDelegate.java
@@ -38,5 +38,5 @@ public interface I_LogDelegate {
      * @param message
      * @param t
      */
-    public void log(int type, Object message, Throwable t);
+    public void log(short type, Object message, Throwable t);
 }
